+++ Examples/RunExamples.cs
@@ -69,6 +69,9 @@ namespace Examples
                     case "ParametrizedQuery":
                         await ParametrizedQuery.Main(args);
                         break;
+                    case "RecordRowExample":
+                        await RecordRowExample.Main();
+                        break;
                 }
             }
             else
@@ -78,7 +81,7 @@ namespace Examples
                                   "FluxClientPocoExample, PlatformExample, WriteEventHandlerExample, WriteApiAsyncExample, " +
                                   "CustomDomainMapping, PocoQueryWriteExample, CustomDomainMappingAndLinq, " +
                                   "SynchronousQuery, InfluxDB18Example, QueryLinqCloud, ManagementExample, " +
-                                  " InvokableScripts, ParametrizedQuery");
+                                  " InvokableScripts, ParametrizedQuery, RecordRowExample");
             }
         }
     }